+++ Hell2025/Hell2025/src/Bvh/Cpu/CpuBvh.cpp
@@ -288,7 +288,7 @@ namespace Bvh::Cpu {
                 int newPrimitiveIndex = static_cast<int>(sceneBvh.m_instances.size());
 
                 // Loop over each primitive instance in the leaf
-                for (int i = 0; i < node.primitiveCount; i++) {
+                for (uint32_t i = 0; i < node.primitiveCount; i++) {
                     int primitiveId = node.firstChildOrPrimitive + i;
                     int instanceIndex = bvh.prim_ids[primitiveId];
 
@@ -318,8 +318,9 @@ namespace Bvh::Cpu {
         g_triangleData.clear();
 
         // Preallocate memory
-        int totalNodeCount = 0;
-        int totalTriangleDataSize = 0;
+        size_t totalNodeCount = 0;
+        size_t totalTriangleDataSize = 0;
+
         for (auto it = g_meshBvhs.begin(); it != g_meshBvhs.end(); ++it) {
             MeshBvh& meshBvh = it->second;
             totalNodeCount += meshBvh.m_nodes.size();
@@ -359,8 +360,8 @@ namespace Bvh::Cpu {
             g_triangleData.insert(g_triangleData.end(), meshBvh.m_triangleData.begin(), meshBvh.m_triangleData.end());
 
             // Increment offsets
-            rootNodeOffset += meshBvh.m_nodes.size();
-            baseTriangleOffset += meshBvh.m_triangleData.size();
+            rootNodeOffset += (uint32_t)meshBvh.m_nodes.size();
+            baseTriangleOffset += (uint32_t)meshBvh.m_triangleData.size();
         }
     }
